ci: post-rebase drift fixes — split async_chkpt per 500-LoC cap, new SsmStatePool bases_dev fields in test pool, clippy is_multiple_of
