+++ crates/spark-model/src/model/ssm_pool.rs
@@ -632,6 +632,12 @@ mod slot_guard_tests {
             conv_intermediate_pools: Vec::new(),
             h_checkpoint_pools: Vec::new(),
             conv_checkpoint_pools: Vec::new(),
+            h_state_bases_dev: DevicePtr::NULL,
+            conv_state_bases_dev: DevicePtr::NULL,
+            h_checkpoint_bases_dev: DevicePtr::NULL,
+            conv_checkpoint_bases_dev: DevicePtr::NULL,
+            h_intermediate_bases_dev: DevicePtr::NULL,
+            conv_intermediate_bases_dev: DevicePtr::NULL,
             h_bytes: 0,
             conv_bytes: 0,
             max_slots,
